+++ TVS-Player/MainWindow.xaml
@@ -48,7 +48,6 @@
             </Grid>
         </StackPanel>
         <Grid Panel.ZIndex="2" x:Name="hideOnClick" Margin="0,0,0,0" Background="#00000000" MouseLeftButtonUp="MenuHide_Click" Visibility="Hidden"/>
-        <Frame x:Name="Frame" Margin="48,0,0,0" NavigationUIVisibility="Hidden" Loaded="FrameLoaded_Handler"/>
-        <Button x:Name="button" Content="WIP Library" HorizontalAlignment="Right" Margin="0,10,10,0" VerticalAlignment="Top" Click="OpenLibrary" Padding="3,1"/>
+        <Frame x:Name="Frame" Margin="48,0,0,0" NavigationUIVisibility="Hidden" Loaded="FrameLoaded_Handler" Source="/TVS-Player;component/Shows.xaml"/>
     </Grid>
 </Window>
