+++ TVS-Player/Shows.xaml
@@ -5,10 +5,41 @@
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
       xmlns:local="clr-namespace:TVS_Player"
       mc:Ignorable="d" 
-      d:DesignHeight="300" d:DesignWidth="300"
+      d:DesignHeight="648" d:DesignWidth="1152"
       Title="Shows">
 
     <Grid>
-        
+        <Grid.RowDefinitions>
+            <RowDefinition Height="60"/>
+            <RowDefinition/>
+            <RowDefinition Height="20"/>
+        </Grid.RowDefinitions>
+        <Canvas >
+            <Canvas.Background>
+                <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
+                    <GradientStop Color="#4C000000" Offset="0"/>
+                    <GradientStop Offset="1" Color="#00000000"/>
+                </LinearGradientBrush>
+            </Canvas.Background>
+        </Canvas>
+        <TextBlock FontSize="20" Text="Knihovna serialu" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10,10,0,0" Foreground="White" />
+        <Button Content="Exit" HorizontalAlignment="Right" VerticalAlignment="Top" Click="Quit_Event"/>
+        <DockPanel x:Name="SearchBarPanel" VerticalAlignment="Bottom" HorizontalAlignment="Right" Height="24" Width="180" Margin="0,0,20,10">
+            <Rectangle Width="24" Margin="0,0,10,0" HorizontalAlignment="Left"/>
+            <TextBox Text="NejakySerial" VerticalContentAlignment="Center" />
+        </DockPanel>
+
+        <ScrollViewer Grid.Row="1">
+            <WrapPanel  x:Name="List" Margin="10,30,10,10">
+                <Grid x:Name="BaseRectangle" Height="200" Width="150" Margin="10">
+                    <Rectangle  Fill="#FF994747" RadiusY="5" RadiusX="5">
+                        <Rectangle.Effect>
+                            <DropShadowEffect BlurRadius="20" Direction="270"/>
+                        </Rectangle.Effect>
+                    </Rectangle>
+                    <TextBlock Text="Random jmeno serialu" Foreground="White" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0,0,0,20"/>
+                </Grid>
+            </WrapPanel>
+        </ScrollViewer>
     </Grid>
 </Page>
